Split Vitest workspace dist config
Move workspace dist resolution and OpenTelemetry SSR policy into focused helpers so the public config wrapper stays readable. No behavior change.

Refs ENG-738
